modify: Transfer codes from HomeActivity to Main
