+++ HealthCareApp/app/src/main/AndroidManifest.xml
@@ -13,13 +13,6 @@
         android:theme="@style/Theme.HealthCareApp"
         tools:targetApi="31">
         <activity
-            android:name=".HomeActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".SignUpActivity"
             android:exported="false">
             <meta-data
